SourceIndexedTypeSystem
Summary:
This diff builds off the source information introduced in D75806744.

In short, a `TypeSystem` implementation *may* optionally support an alternative lookup scheme based on *names*. If it does, it should return a non-null pointer from `TypeSystem::asSourceIndexed()`.

The `SourceIndexedTypeSystem` offers two capabilities:
1. Look up a type directly by `(location, name)`.
2. List all types within one `location`.

The return values are all `DefinitionRef` meaning the underlying structures are owned by the originating `TypeSystem`. In other words, the data is the exact same — we just offer an alternative way to index their (non-unique) identities.

Reviewed By: iahs, pranavtbhat

Differential Revision: D75806743

fbshipit-source-id: 67b2ce83cb2476c8c49be8c0990ea307ff181aeb