Dynamic cursor serialization
Summary:
A version of the existing cursor serialization API that does not require static schema.
Initial implementation doesn't support dynamic schema either, and enforces compact/binary protocol as a result, but the ability to associate a type_system::TypeRef at the root level is planned.
This is intended to be used internally to implement public operation APIs rather than directly by users, so it is in the detail namespace.
Two example operations are implemented and benchmarked in D77672020 to demonstrate usage; they will be productionized in future diffs.

Reviewed By: thedavekwon, praihan

Differential Revision: D77606042

fbshipit-source-id: e77f1c5ba59fb1aa4c0eebce15f8edabe6880733