simplify async processor factory
Summary:
Rather than have `PythonAsyncProcessor` store a map of 3 `PythonMetadata`, just create them as needed. The `PythonMetadata` only has enum elements, so it's cheap to create.

Reasoning: the current logic is hard to follow. Also, the map lookups are not free. By removing the map, we mildly decrease the memory footprint. The cost of allocating the `shared_ptr` is offset by removing the map lookup.

Reviewed By: prakashgayasen

Differential Revision: D77956737

fbshipit-source-id: e6a04eeee76f7cdb68e6a157352ce765d5d46b74