Fixture to show fix for external program metadata for services
Summary:
The fix for the next diff in this stack, {D77617349}, adds an import that corresponds to an included thrift file (a.k.a external program) for a service from an included thrift file that another service extends.

In order to see this include change in the fixtures, add a fixture that reproduces this use case.

Reviewed By: ahilger

Differential Revision: D77617336

fbshipit-source-id: c743024c0f94b5fb55b2298970f34b2da7e471ae