Move testing.thrift and deps to thrift/lib/python/test
Summary:
A bunch of stuff in `python/test` depends on `testing.thrift`, so it should live in `python/test`. The alternative is making some duplicate version of `testing.thrift` and migrating everything in `thrift/lib/python` to it :/

Luckily it's not too hard to update because `namespace py3` is consistent -> no `.py` file changes.

Reviewed By: createdbysk

Differential Revision: D78188480

fbshipit-source-id: 52450b7b3f14bed915a2852a51199ef7b7849c10