Add tests for metadata for extension of services with no types
Summary:
# Purpose of this test
If you stack this diff on top of the changes in D77617854, then

```
buck2 test fbcode//thrift/lib/python/test:metadata
```
will fail with P1867281226

```
NameError: name '_fbthrift__testing__base_service_only__thrift_metadata' is not defined`
```

That matches the error a user reported in [this post](https://fb.workplace.com/groups/thrift.py3/permalink/3173980906087013/).

Like the user mentioned, D77617349 resolves this error.

Had this test existed before this regression, we would have caught this error and not let the regression happen with  D77617854.

# How did this happen?
In the original chronology, the changes in D77617349 were before D77617854 in the stack. I changed the order because that seemed to make the progression clearer between the two changes. Due to the lack of the test the current diff adds, the breakage that happened between those two diffs went unnoticed.

Reviewed By: ahilger

Differential Revision: D78185189

fbshipit-source-id: a57183eb3d7c0b2ad8d50b3c1d7ed8faafab7a16