fastpath first pass
Summary:
This diff comes from this [write up](https://fb.workplace.com/permalink.php?story_fbid=750706237302375&id=100070890163485)

This fast path implementation attempts to deserialize a struct's fields in order. If it comes across a field that is out of order it falls back to the previous logic that used a `match` statement inside of a `loop`

The new logic [looks like this](https://www.internalfb.com/phabricator/paste/view/P1863984154?lines=220-282)

It is important to note that there is no attempt to re-enter the fast path after a miss. Some scenarios might not see a large improvement because of this.

Reviewed By: slawlor

Differential Revision: D77982563

fbshipit-source-id: 1bb7f404bb30bd0b7ce76bc0654ca143aa274d0f