Make caching configurable
Summary:
This diff introduces a new feature to make struct field caching configurable. By adding the `python.DisableFieldCache` annotation on a struct, users can now opt-out of all fields caching for specific structs.

The primary motivation behind this change is to improve performance in `First/Uncached` scenarios. This is particularly useful when a struct is used as a field in another struct.

**Benefits:**
1. Improved performance in First/Uncached scenarios
2. With this update, users have more control over how field caching is applied, allowing them to fine-tune their Thrift configurations for optimal performance.

**How to disable field caching?**
To disable field caching for a struct, simply add the python.DisableFieldCache annotation above the struct definition.

```
python.DisableFieldCache
struct MyStruct {
  1: i32 id,
  2: string name,
}
```

Reviewed By: ahilger

Differential Revision: D77962469

fbshipit-source-id: a394391c120eb58e07429c7e9c90c059dc03941f