implement key sorted serialize: sets
Summary:
This ended up being fairly tricky for a couple reasons.
1. the `types/typeinfo` sub-template is invoked in a lot of places where field context is not available. Also, didn't want to change the `__init__` for `SetTypeInfo`, because key sorting is only invoked in the context of fields. The lead me to add the builder-pattern-style `enableKeySorted()` method that returns the `SetTypeInfo` with the key sorting toggled on in cpp layer.
2. The table-based-serializer needs function pointers to `static` methods, so serializability cannot be an instance attribute, only a template parameter.

Given the above two considerations, it made sense to keep a single python class `SetTypeInfo`, but its `cpp_obj` needs to be a pointer that accommodates both the default and key-sorted types. Hence added a `Base` a template-less abstract base class, and made the template instantiation a `final` child, which should force de-virtualization of the couple virtual methods.

Reviewed By: prakashgayasen

Differential Revision: D78382663

fbshipit-source-id: 6acdbfc05af7929cf71dc3001acfa51a5ea53644