Introduce asSyntaxGraph*
Summary:
When TypeSystemFacade is utilized to convert SyntaxGraph to TypeSystem, we provide `asSyntaxGraph*` methods to get the SyntaxGraph node back from TypeSystem node.

This will be useful to fetch the annotation when using TypeSystem via SyntaxGraph.

Context:
In schema.thrift, we store annotations as Protocol Value similar to JSON. In SyntaxGraph, we are currently serving these annotations via folly::dynamic. For type_system.thrift, ideally we should be serving annotations as DynamicValue, which is not currently available. In the mean time, users can leverage this method to convert back to SyntaxGraph Node to fetch the annotation.

Reviewed By: pranavtbhat, praihan

Differential Revision: D78458906

fbshipit-source-id: 796909aa66b97cccd4330efe0e4aa371a5f9bdb6