pass in server name to makeRpcMetadataUtil for Scuba logging
Summary:
## Server Name annotation
When we make an RPC call, makeRpcMetadataUtil only supplies methodName so I am modifying the API to pass in serverName as well

## Scuba logging for contextprop client interceptor rollout

We'd like to have better observability into shift from old code path to the interceptorFrameworkMetadata path initialized by ContextProp Client Interceptor.

With this Scuba log, we can ensure that isFromInterceptor bool shifts from 0 to 1 for a service once we install to that service - with roughly the same distribution of ThriftFrameworkMetadata sizes (ideas here welcome - i'd expect distribution of framework_metadata_size to be the same). Basically ensure that pre and post rollout the TFM characteristics shouldn't change except the code path

Reviewed By: praihan

Differential Revision: D78106429

fbshipit-source-id: 43efedc4d67989583ebf7438632564fda9a84513