fix converter bug; use immutable serializer
Summary:
Under certain circumstances, mutations to structs in containers of a parent struct are dropped when the parent struct is serialized, even though the mutations appear to have succeeded when inspected via normal Python field access.

As a temporary mitigation, convert to immutable python then use its serializer.

Note, this diff also fixes a pre-existing bug in the converter w/r/t fields that are MutableGeneratedError. When the rest of diff is reverted, the converter.pyx change should be retained.

Reviewed By: aristidisp, prakashgayasen

Differential Revision: D78831902

fbshipit-source-id: 0baf133e61ff13ee2ebe70cb196bc594fe3eec48