generalize handling of stream
Summary:
The generated `.py` client code gives the async_client `_send_request` method a `response_cls`. For normal request-response, this is the response class itself. However, for streams it is a tuple. As we implement sink and bidi-streaming, there will be other `tuple` `response_cls` types, so create a more generalized approach to get the `RpcKind`.

We can't deduce from the `tuple` length because the bidi stream will have tuple length 2, just like regular stream. Also, this may change in the future if a first response is needed, if a final response type is needed, etc. Therefore, encode the rpc kind directly as a class variable on the streaming class, for stream, sink, and bidi.

Reviewed By: prakashgayasen

Differential Revision: D79025308

fbshipit-source-id: 23801b091d467f1fbdbed1756d3acfa324fc3968