Prevent punctuation from being absorbed into string delimiters in bundling logic
Summary:
The change to RpcMetadata.thrift, which is a bundled file, puts one of the chunk delimiters at a closing bracket:
```
  MORE = 2, // More compression/CPU/memory.)__FBTHRIFT_TAG__"
R"__FBTHRIFT_TAG__(}
```

This closing bracket is interpreted as part of the raw string delimiters:

```
R"d-char-seq ﻿(optional)(r-char-seq ﻿(optional))d-char-seq ﻿(optional)"

d-char-seq	-	A sequence of one or more d-char ﻿s, at most 16 characters long
d-char	-	A character from the basic source character set(until C++23)basic character set(since C++23), except parentheses, backslash and spaces
```

This produces the unhelpful error
```
[ERROR:thrift/lib/thrift/RpcMetadata.thrift:152] expected identifier
```

To prevent this, adjust the chunking logic to seek forward until a line that starts with a space or line break. These are not d-chars so will not be absorbed into the delimiter.

Reviewed By: sazonovkirill

Differential Revision: D79209027

fbshipit-source-id: d65791fff661c27b7958ed6d60b5addf919f9e1f