Remove explicit self. property navigation from C++ Whisker templates
Summary:
Remove `self.` in C++ templates for consistency and to improve readabiity.

In D79021920, we added an implicit lookup into the `self.` Whisker prototype in the `mstch_compat` layer, as a fallback to mstch properties.
In D79104932, we cleaned up mstch definitions that may take precedence over the Whisker prototype properties, requiring an explicit `self.` navigation.

Reviewed By: iahs

Differential Revision: D79277424

fbshipit-source-id: b58e52926abc0255696d19f84762d2155c3b1cd7