don't segfault if used outside of async context manager
Summary:
Currently, it segfaults in opt at the beginning of `_send_request_inner` when `self._omni_client` is deref'd. One mitigation would be to `await self._connect_future`, but this creates future problems.

We want to encourage customers to migrate to `async with` context manager usage of client, in order to ensure `_atexit_callbacks` are handled properly with `__aexit__`.

Stored clients are problematic because:
1. _atexit_callbacks` can only be handled properly with `__aexit__`. Can't safely run an `async` callback at `__del__` time.
2. A stored client can be invalidated if used in async context manager, then fail unexpectedly. If a customer *really* wants to store a context manager, perhaps they can take responsibility for invoking `__aenter__` and `__aexit__` manually.

Reviewed By: prakashgayasen

Differential Revision: D79293616

fbshipit-source-id: 7b748e80ebb6f396155c79253b4997fa294b471e