Add thrift server event to make resource pools stragglers easier to track.
Summary:
This event should help us track hold-out services that aren't using resource pools.

There were some problems with trying to track hold-outs via the "serve" event. Mostly, the sampling rate of 10000 makes it possible for some services to slip through the cracks.

By adding this new event and setting the sampling rate to 1 in the previous diff, we'll be able to reliably track services that don't use resource pools.

Reviewed By: sazonovkirill

Differential Revision: D79379270

fbshipit-source-id: 17324ab0fbc4d5b741486bc7ffd4ae27be56ed77