Add callbacks for BiDi request type
Summary:
The bi-directional stream is being modeled as two independent streams in opposite directions that share a first response. The two half-streams are called Stream and Sink and have symmetric methodsets across the two callbacks.
A notable departure from the existing stream/sink contracts is that error/complete/cancel methods only terminate one of the half-streams, requiring two lifetime bools to be returned instead of one (encapsulated in the `BiDiChannelState` struct) from both terminating and non-terminating methods. (This supports the case where the implementation of a method for one half-stream calls a termiating method on the other half-stream).

Reviewed By: sazonovkirill

Differential Revision: D79598718

fbshipit-source-id: 979b927efa95209fe666ea82ab1afe4bd853b20f