Add extra-validation: required_field_qualifier
Summary:
This change introduces the `required_field_qualifier` extra validation option, which controls the level of diagnostic (none/warn/error) for fields that are marked as `required`, but not annotated with the newly-introduced `thrift.AllowUnsafeRequiredFieldQualifier`.

The default behavior is unchanged: a warning will be issued.

We expect to switch the default behavior to `error` when all existing uses have been annotated.

Reviewed By: iahs

Differential Revision: D80068264

fbshipit-source-id: f605684c894157e75bd2103b923e2fc7f199576e