Implement the transform to add namespace py3 to a thrift file.
Summary:
# What
* Implement the transform that adds `namespace py3` to the given Thrift file.
* Add additional arguments to the transformer to allow this implementation to pass the namespace py3 as an additional argument.

# Why does transformer need the additional arguments?
Step 1 of the codemod to Enable thrift-python is to add `namespace py3` to the thrift file. The `namespace py3` is not based only on the contents and location of the thrift file. See below for the algorithm. All the Thrift codemods so far relied on the contents of the Thrift file alone. This new codemod relies on a value, the `namespace py3` name, outside of a Thrift file.

# Algorithm to determine the `namespace py3`
The process that generates this diff uses the value of the `py_base_module` to determine the value of `py3_namespace`. If this diff adds a `namespace py3` to thrift file(s), then those values will match the `py3_namespace`.
* If the `py_base_module` is a non-empty string, then `py3_namespace` will match the `py_base_module`.
* If the `py_base_module` is an empty string or not present, then `py3_namespace` will be the dotted path to the buildfile (`BUCK` or `TARGETS`) that contains the `thrift_library`.

Reviewed By: vitaut

Differential Revision: D80091252

fbshipit-source-id: ff47633c6a9a9e9cd021bd76f973b27227181981