Add built-in ternary if function
Summary:
Add built-in ternary function `if cond true_val false_val`. This allows users to express ternary values, e.g. `let` assignments that differ based on a condition.

The conditional block syntax `{{#if cond}}{{#let foo = true_val}}{{#else}}{{#let foo = false_val}}` is not useful for this scenario, since the `foo` variable will only exist in the scope of the `if`/`else` blocks.

It is implemented as a built-in rather than in the Whisker standard library so that the expression evaluation can be short-circuited to only the matching branch, as with other logical built-in functions.

Reviewed By: vitaut, praihan

Differential Revision: D80291640

fbshipit-source-id: ce142d7479be50c2599fd7c58a3b3e7fa219a3d9